Move select_rows to utils
